feat: Update DropdownButton
BREAKING CHANGE: We add `textVariant`, `spaceBetween` and `disabled` props to
be able to change the size of the icon, and disabled the component.
Besides it's no longer needed to use Typography as first child of
DropdownButton since it's wrapped inside, but the old school way
still works as before.
It's still possible to use the Dropdown without the button wrapper, as before, use `Dropdown` component instead of `DropdownButton`.